[CH##160914136] Restructured the code in readyness to host in Heroku
