--- v3/__init__.py
@@ -1,7 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-
-
-
